# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ONDC Integration API",
    "version": "1.0.0",
...
    "contact": {
      "name": "Vaatun Technologies",
      "email": "vaibhav@vaatun.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    },
    {
      "url": "https://ondc-staging.vaatun.com",
      "description": "Staging environment"
    }
  ],
  "components": {
    "schemas": {
...
      "SelectResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "object",
            "properties": {
              "ack": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "string",
                    "example": "ACK"
                  }
                },
                "required": [
                  "status"
                ]
              }
            },
            "required": [
              "ack"
            ]
          },
          "transactionId": {
            "type": "string",
            "format": "uuid",
            "example": "019abc12-3456-7890-abcd-ef1234567890"
          },
          "messageId": {
            "type": "string",
            "format": "uuid",
            "example": "019abc12-3456-7890-abcd-ef1234567891"
          }
        },
        "required": [
          "transactionId",
          "messageId"
        ],
        "additionalProperties": {}
      },
      "SelectAddOn": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "addon-001"
          },
          "quantity": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "example": 1
          }
        },
        "required": [
          "id",
          "quantity"
        ]
      },
      "SelectRequest": {
        "type": "object",
        "properties": {
          "transactionId": {
            "type": "string",
            "format": "uuid",
            "example": "019abc12-3456-7890-abcd-ef1234567890"
          },
          "bppId": {
            "type": "string",
            "example": "ondc-mock-server.example.com"
          },
          "bppUri": {
            "type": "string",
            "format": "uri",
            "example": "https://ondc-mock-server.example.com/api/ondc"
          },
          "providerId": {
            "type": "string",
            "example": "provider-001"
          },
          "itemId": {
            "type": "string",
            "example": "item-001"
          },
          "parentItemId": {
            "type": "string",
            "example": "item-001"
          },
          "xinputFormId": {
            "type": "string",
            "example": "form-001"
          },
          "xinputSubmissionId": {
            "type": "string",
            "example": "submission-001"
          },
          "addOns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SelectAddOn"
            },
            "example": [
              {
                "id": "addon-001",
                "quantity": 1
              }
            ]
          }
        },
        "required": [
          "transactionId",
          "bppId",
          "bppUri",
          "providerId",
          "itemId",
          "parentItemId"
        ]
      },
...
      "Quote": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "quote-001"
          },
          "price": {
            "type": "object",
            "properties": {
              "currency": {
                "type": "string",
                "example": "INR"
              },
              "value": {
                "type": "string",
                "example": "15000.00"
              }
            },
            "required": [
              "currency",
              "value"
            ],
            "additionalProperties": {}
          },
          "breakup": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "Base Premium"
                },
                "price": {
                  "type": "object",
                  "properties": {
                    "currency": {
                      "type": "string",
                      "example": "INR"
                    },
                    "value": {
                      "type": "string",
                      "example": "15000.00"
                    }
                  },
                  "required": [
                    "currency",
                    "value"
                  ],
                  "additionalProperties": {}
                },
                "item": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "add_ons": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          }
                        },
                        "required": [
                          "id"
                        ]
                      }
                    }
                  },
                  "additionalProperties": {}
                }
              },
              "required": [
                "title",
                "price"
              ],
              "additionalProperties": {}
            }
          },
          "ttl": {
            "type": "string",
            "example": "PT30M"
          }
        },
        "required": [
          "price"
        ],
        "additionalProperties": {}
      },
      "SelectProvider": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "provider-001"
          },
          "descriptor": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "ICICI Prudential"
              },
              "short_desc": {
                "type": "string"
              },
              "long_desc": {
                "type": "string"
              },
              "images": {
                "type": "array",
                "items": {}
              }
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "id"
        ],
        "additionalProperties": {}
      },
      "SelectItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "item-001"
          },
          "parent_item_id": {
            "type": "string"
          },
          "descriptor": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "example": "Health Guard Plus"
              },
              "short_desc": {
                "type": "string"
              },
              "long_desc": {
                "type": "string"
              },
              "images": {
                "type": "array",
                "items": {}
              }
            },
            "additionalProperties": {}
          },
          "price": {
            "type": "object",
            "properties": {
              "currency": {
                "type": "string",
                "example": "INR"
              },
              "value": {
                "type": "string",
                "example": "15000.00"
              }
            },
            "required": [
              "currency",
              "value"
            ],
            "additionalProperties": {}
          },
          "xinput": {},
          "add_ons": {
            "type": "array",
            "items": {}
          },
          "tags": {
            "type": "array",
            "items": {}
          }
        },
        "required": [
          "id"
        ],
        "additionalProperties": {}
      },
      "SelectResultsResponse": {
        "type": "object",
        "properties": {
          "found": {
            "type": "boolean",
            "example": true
          },
          "transactionId": {
            "type": "string",
            "example": "019abc12-3456-7890-abcd-ef1234567890"
          },
          "messageId": {
            "type": "string",
            "example": "019abc12-3456-7890-abcd-ef1234567891"
          },
          "itemId": {
            "type": "string",
            "example": "item-001"
          },
          "providerId": {
            "type": "string",
            "example": "provider-001"
          },
          "hasResponse": {
            "type": "boolean",
            "example": true
          },
          "quote": {
            "$ref": "#/components/schemas/Quote"
          },
          "provider": {
            "$ref": "#/components/schemas/SelectProvider"
          },
          "item": {
            "$ref": "#/components/schemas/SelectItem"
          },
          "xinput": {
            "description": "XInput form details for additional data collection"
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "additionalProperties": {}
          },
          "message": {
            "type": "string",
            "example": "No select entry found for this transaction"
          }
        },
        "required": [
          "found",
          "transactionId",
          "messageId",
          "hasResponse"
        ],
        "additionalProperties": {}
      }
    },
...
    "securitySchemes": {
      "Ed25519Signature": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "\n# ONDC Ed25519 Signature Authentication\n\nAll outbound requests to ONDC Registry and Gateway require Ed25519 digital signatures using HTTP Signature standard.\n\n## Authentication Flow\n\n### Step 1: Calculate BLAKE-512 Digest\n```typescript\nimport { createHash } from \"crypto\";\n\nconst requestBody = JSON.stringify(payload);\nconst digest = createHash(\"blake2b512\")\n  .update(requestBody)\n  .digest(\"base64\");\n```\n\n### Step 2: Create Signing String\n```typescript\nconst created = Math.floor(Date.now() / 1000) - 3; // Clock drift compensation\nconst expires = created + 300; // 5 minute validity\n\nconst signingString =\n  `(created): ${created}\\n` +\n  `(expires): ${expires}\\n` +\n  `digest: BLAKE-512=${digest}`;\n```\n\n### Step 3: Sign with Ed25519\n```typescript\nimport sodium from \"libsodium-wrappers\";\n\nawait sodium.ready;\n\nconst privateKey = sodium.from_base64(process.env.SIGNING_PRIVATE_KEY);\nconst signature = sodium.crypto_sign_detached(signingString, privateKey);\nconst signatureBase64 = sodium.to_base64(signature);\n```\n\n### Step 4: Construct Authorization Header\n```typescript\nconst keyId = `${subscriberId}|${uniqueKeyId}|ed25519`;\n\nconst authHeader =\n  `Signature keyId=\"${keyId}\",` +\n  `algorithm=\"ed25519\",` +\n  `created=\"${created}\",` +\n  `expires=\"${expires}\",` +\n  `headers=\"(created) (expires) digest\",` +\n  `signature=\"${signatureBase64}\"`;\n```\n\n## Clock Drift Compensation\n\nThe implementation subtracts 3 seconds from the current timestamp to account for clock drift between your server and ONDC servers:\n\n```typescript\nconst created = Math.floor(Date.now() / 1000) - 3; // Subtract 3 seconds\n```\n\nThis prevents authentication failures due to slight time differences.\n\n## Key Management\n\n### Environment Variables Required\n\n- `SIGNING_PRIVATE_KEY`: Ed25519 private key (base64-encoded, 64 bytes)\n- `UNIQUE_KEY_ID`: Unique identifier for this key pair (e.g., \"custom-key-id\")\n- `SUBSCRIBER_ID`: Your registered ONDC subscriber ID (e.g., \"ondc-staging.vaatun.com\")\n\n## Implementation Reference\n\nSee full implementation in:\n- `/src/lib/ondc/client.ts` - ONDCClient.createAuthorizationHeader()\n- `/src/lib/ondc/signing.ts` - Ed25519 signing utilities\n- `/src/entities/tenant.ts` - Key management\n\n## Inbound vs Outbound Authentication\n\n### Outbound (Your App → ONDC)\n- **Routes**: `/api/ondc/lookup`, `/api/ondc/subscribe`, `/api/ondc/search`, `/api/ondc/select`\n- **You sign**: Using your `SIGNING_PRIVATE_KEY`\n- **ONDC verifies**: Using your public key from registry\n\n### Inbound (ONDC → Your App)\n- **Routes**: `/api/ondc/on_subscribe`, `/api/ondc/on_search`, `/api/ondc/on_select`\n- **ONDC signs**: Using their private key\n- **You verify**: Using their public key from registry (retrieved via `/lookup`)\n    "
      }
    }
  },
  "paths": {
...
    "/api/ondc/select": {
      "post": {
        "summary": "Select Insurance Item",
        "description": "Select a specific insurance item to get a quote.\n\nThis endpoint sends a select request to the BPP (seller) to get pricing details for a specific item.\n\n**Flow:**\n1. Receives item selection with provider and item IDs\n2. Optionally includes XInput form submission for eligibility\n3. Sends signed request directly to BPP\n4. Returns transaction ID to poll for quote via on_select callback",
        "tags": [
          "Gateway"
        ],
        "operationId": "select",
        "directoryConfig": {
          "title": "Select Item",
          "description": "Select an insurance item to get a quote",
          "payload": {
            "transactionId": "019abc12-3456-7890-abcd-ef1234567890",
            "bppId": "ondc-mock-server.example.com",
            "bppUri": "https://ondc-mock-server.example.com/api/ondc",
            "providerId": "provider-001",
            "itemId": "item-001",
            "parentItemId": "item-001"
          }
        },
        "requestBody": {
          "required": true,
          "description": "Item selection details",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SelectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Select request accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields"
          },
          "503": {
            "description": "Service unavailable"
          }
        }
      }
    },
...
    "/api/ondc/select-results": {
      "get": {
        "summary": "Poll Quote Results",
        "description": "Polling endpoint to fetch quote (select) results for a transaction.\n\nCalled by the frontend to retrieve on_select response with pricing details.\n\n**Flow:**\n1. Client polls with transaction_id and message_id from initial select\n2. Returns quote with price breakup when available\n3. May include XInput form for additional data collection\n4. Poll until response received or timeout",
        "tags": [
          "Internal"
        ],
        "operationId": "getSelectResults",
        "directoryConfig": {
          "title": "Poll Quote Results",
          "description": "Fetch quote results by transaction and message ID"
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "019abc12-3456-7890-abcd-ef1234567890",
              "description": "Transaction ID from the initial select request"
            },
            "required": true,
            "description": "Transaction ID from the initial select request",
            "name": "transaction_id",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "example": "019abc12-3456-7890-abcd-ef1234567891",
              "description": "Message ID from the initial select request"
            },
            "required": true,
            "description": "Message ID from the initial select request",
            "name": "message_id",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Select results (may be empty if no response yet)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SelectResultsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Missing transaction_id or message_id query parameters"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
...
}